finalizing FE for user

+++ src/main/java/com/b2/reservation/request/KuponRequest.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class KuponRequest {
     private String name;
-    private Double percentageDiscounted;
+    private Integer percentageDiscounted;
 }
